Fix bug in pad converter. input_rank maybe 0 and will set wrong t_pads.sizes().

+++ onnx/version_converter/adapters/pad_10_11.h
@@ -21,7 +21,7 @@ class Pad_10_11 final : public Adapter {
       for (int64_t shape : node->is(kpads)) {
         data_pads.emplace_back(shape);
       }
-      t_pads.sizes() = std::vector<int64_t> {data_pads.size()};
+      t_pads.sizes() = std::vector<int64_t> {(int64_t)data_pads.size()};
       Value* v_pads = graph->addInitializerAndInput(t_pads);
       node->addInput(v_pads);
       node->removeAttribute(kpads);
